Make cli configuration available
