+++ dia-batching-server/Cargo.toml
@@ -28,8 +28,6 @@ smol_str = { version = "0.1.18", features = ["serde"] }
 
 tokio = { version = "0.2.25", features = ["full"] }
 
-structopt = "0.3.25"
-
 rust_decimal = "1.17.0"
 rust_decimal_macros = "1.17"
 futures = "0.3.30"
